0.2.7-buildfix3
- added, fixed and tested msys64 build mode
- icon for x64
- libfreeglut no longer needs to be included in the repo

+++ tools/cmakeautomate_msys64.py
@@ -26,7 +26,6 @@ def zip_build(source_dir, output_zip):
                 arcname = os.path.relpath(file_path, source_dir)
                 zipf.write(file_path, arcname)
 
-
 # Get the current script's directory
 script_dir = os.path.dirname(os.path.abspath(__file__))
 
@@ -42,40 +41,44 @@ folder_name = "build"
 build_folder_path = os.path.join(script_dir, folder_name)
 
 if os.path.exists(build_folder_path):
-    shutil.rmtree(build_folder_path)
+    # shutil.rmtree(build_folder_path)
+    for root, dirs, files in os.walk(build_folder_path, topdown=False):
+        for name in dirs:
+            shutil.rmtree(os.path.join(root, name))
+
+        for name in files:
+            if name != 'ninja.exe':
+                os.remove(os.path.join(root, name))
+    print("build directory cleaned.")
 
 os.makedirs(build_folder_path, exist_ok=True)
 
 # Define the URL to download and the filename
 url = "https://github.com/ninja-build/ninja/releases/download/v1.11.1/ninja-win.zip"
 filename = "ninja-win.zip"
+
 ninja_zip_path = os.path.join(build_folder_path, filename)
+ninja_exe_path = os.path.join(build_folder_path, "ninja.exe")
 
 unzip_command = f"powershell Expand-Archive -Path {ninja_zip_path} -Destination {build_folder_path}"
-
-# Download the file using curl
 download_command = f"curl -L -o {ninja_zip_path} {url}"
-subprocess.call(download_command, shell=True)
 
-# Check if the download was successful
-if os.path.exists(ninja_zip_path):
-    # Extract the downloaded zip file
-    subprocess.call(unzip_command, shell=True)
+if not os.path.exists(ninja_exe_path):
+    # Download the file using curl
+    subprocess.call(download_command, shell=True)
 
-    # Clean up the zip file
-    os.remove(ninja_zip_path)
-    print("Done downloading ninja.")
-else:
-    print("Download failed.")
-    exit()
+    # Check if the download was successful
+    if os.path.exists(ninja_zip_path):
+        # Extract the downloaded zip file
+        subprocess.call(unzip_command, shell=True)
 
-ninja_exe_path = os.path.join(build_folder_path, "ninja.exe")
-mingw_path = os.path.join(script_dir, "mingw32")
-# if os.path.exists(mingw_path):
-#     c_compiler_path = os.path.join(script_dir, r"mingw32\bin\gcc.exe").replace("\\","/")
-#     cxx_compiler_path = os.path.join(script_dir, r"mingw32\bin\c++.exe").replace("\\","/")
-#     subprocess.call(f"cmake -G \"Ninja\" -DCMAKE_MAKE_PROGRAM={ninja_exe_path} -DCMAKE_PREFIX_PATH={mingw_path} -DCMAKE_C_COMPILER={c_compiler_path} -DCMAKE_CXX_COMPILER={cxx_compiler_path} ..", shell=True, cwd=build_folder_path)
-# else:
+        # Clean up the zip file
+        os.remove(ninja_zip_path)
+        print("Done downloading ninja.")
+    else:
+        print("Download failed.")
+        exit()
+
 subprocess.call("cmake -G \"Ninja\" -DCMAKE_MAKE_PROGRAM=ninja.exe ..", shell=True, cwd=build_folder_path)
     
 subprocess.call("cmake ..", shell=True, cwd=build_folder_path)
